requirements.txt
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
opencv-python
jupyter
pillow